+++ src/components/detailedCard/style.module.css
@@ -7,6 +7,7 @@
   padding: 0 14px 14px;
   margin: 20px 0;
   font-family: Arial, sans-serif;
+  color: black;
   background-color: #f9f9f9;
   box-shadow: 0 4px 10px rgba(0, 0, 0, 0.3);
   border-radius: 10px;
