+++ src/hooks/useAppDispatch.ts
@@ -1,4 +1,4 @@
 import { useDispatch } from 'react-redux';
-import { AppDispatch } from '../api/store.ts';
+import { AppDispatch } from '../api/store';
 
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
