+++ src/hooks/useAppSelector.ts
@@ -1,4 +1,4 @@
 import { TypedUseSelectorHook, useSelector } from 'react-redux';
-import { RootState } from '../api/store.ts';
+import { RootState } from '../api/store';
 
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
