+++ src/index.css
@@ -4,7 +4,7 @@
   font-weight: 400;
 
   color-scheme: light dark;
-  color: black;
+  color: white;
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
